backend/requirements.txt
# FastAPI & Server
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
websockets>=12.0

# Document Processing
pdfplumber==0.11.4
Pillow>=10.0.0

# Vector Store & Embeddings
qdrant-client==1.12.1
mistralai==1.2.0

# ElevenLabs
elevenlabs==1.10.0

# Web Search
duckduckgo-search==6.3.5

# Evaluation
wandb>=0.18.0

# Utils
requests>=2.28.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
